requirements.txt
gradio>=4.13.0
openai>=1.1.0
python-dotenv>=1.0.0
langchain>=0.3.0
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
pypdf>=3.15.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.28.0 